refactor: move cmap tests into a dedicated file
